feat: add partial support for Nordea
This patch refactors OsuuspankkiCertService to its own file (so that we
don't bloat the pankkiyhteys.ts file too much, and then adds classes
similar to OsuuspankkiCertService and Osuuspankki (Client), but for
Nordea.

https://www.nordea.fi/en/business/our-services/mobile-online-services/web-services.html

This does NOT yet implement the certificate request/renewal functions
for Nordea. The user must obtain the Nordea certificate themselves,
for example with the Nordea Security Client that is mentioned in the Web
Services document.

NOTE: Nordea support required adding some new properties to the request
methods. For example, the "Command" parameter is always required by
Nordea. As far as I know this should not break Osuuspankki support, but
that'll have to be verified to avoid regressions.